fix(ai): extract readFileContent so its unit test skips the utils barrel
The readFileContent test imported the agent/utils barrel, which (since the
OAuth merge added oauthPersist -> core/store) transitively pulls
agent/Agent -> googleAuth (LWC @api decorators) into the graph, crashing the
--experimental-strip-types loader (the failing `unit` job on the 2.0.6 RC).

Move readFileContent to its own named module and re-export it from the barrel
(app.ts keeps importing from agent/utils). The test now imports the standalone
module directly, so it never loads core/store. This avoids the earlier broken
attempt of deep-importing oauthPersist via a 3-segment specifier the LWC
bundler can't resolve — the barrel and all module specifiers are unchanged, so
the extension bundle is unaffected.
